+++ scratch.js
@@ -22,7 +22,21 @@ nlp.extend(require('./plugins/sentences/src'))
 // console.log(nlp('next week').dates().json())
 // nlp('you are John, Lisa, Fred').match('#FirstName{1,2}').debug()
 
-let doc = nlp('i may really drive').debug()
+let doc = nlp('i seriously could even drive').debug()
 doc.verbs().toParticiple()
 console.log(doc.text())
 // i thought he really could have.
+
+// let doc = nlp('i did really walk')
+// doc.sentences().toPresentTense()
+// console.log(doc.text())
+
+/*
+    Snowboarding is a winter sport.   [snowboarding = subject]
+    I love snowboarding.    [snowboarding = object ]
+    I am excited by snowboarding.   [snowboarding = object of a preposition]
+    One popular sport is snowboarding.   [snowboarding = subject complement]
+    Right now, the athlete is snowboarding.   [is snowboarding = present continuous]    
+    He was snowboarding yesterday afternoon.   [was snowboarding = past continuous]
+    Tomorrow, my friends and I are going to be snowboarding.   [are going to be snowboarding = future 
+*/