+++ changelog.md
@@ -8,9 +8,11 @@ compromise uses semver, and pushes to npm frequently
 
 While all _Major_ releases should be reviewed, our only two _large_ releases are **v6** in 2016 and and **v12** in 2019. Others have been mostly incremental, or niche.
 
-#### [Unreleased]
+#### 13.4.0 [Unreleased]
 
-- add `verbs().toParticiple()` method  (add to observables/verb)
+- **[change]** stop including adverbs in `.conjugate()` results
+- **[new]** add `verbs().toParticiple()` method  (add to observables/verb)
+- **[new]** add `sentences().toParticiple()` method  (add to observables/verb)
 
 #### 13.3.2